Finished TreeNodes Started reading Journal

+++ Database/Entities/Entity.cs
@@ -1,10 +1,6 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-
 namespace Database;
 
 public class Entity
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
 }
